+++ frontend/src/components/NavComponent.js
